refactor: extract SubsonicBrowsingController — fixes #71
